refactor: Remove prisma ORM
