+++ app/routes/blog+/$.tsx
@@ -18,7 +18,7 @@ export function ErrorBoundary() {
         404: () => (
           <div className="prose flex flex-col gap-6 dark:prose-invert">
             <h1 className="mb-0">We can&apos;t find this post 😅</h1>
-            <Link to="/md-blog" className="flex items-center text-lg underline">
+            <Link to="/blog" className="flex items-center text-lg underline">
               <ArrowLeft />
               <span>Back to posts</span>
             </Link>
