+++ package.json
@@ -8,7 +8,7 @@
     "build:server": "tsx ./other/build-server.ts",
     "dev": "tsx watch --clear-screen=false --ignore '.cache/**' --ignore 'app/**' --ignore 'vite.config.ts.timestamp-*' --ignore 'build/**' --ignore 'node_modules/**' --inspect ./server/index.ts",
     "start": "cross-env NODE_ENV=production node ./server-build/index.js",
-    "setup": "run build && run build:server && prisma generate && prisma migrate deploy && prisma db seed",
+    "setup": "run build && run build:server",
     "lint": "eslint --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
     "format": "prettier --write .",
     "typecheck": "tsc"
@@ -19,7 +19,6 @@
     "@epic-web/client-hints": "^1.3.2",
     "@epic-web/invariant": "^1.0.0",
     "@mdx-js/rollup": "^3.0.1",
-    "@prisma/client": "^5.16.2",
     "@radix-ui/colors": "^3.0.0",
     "@radix-ui/react-dropdown-menu": "^2.1.1",
     "@radix-ui/react-navigation-menu": "^1.2.0",
@@ -81,7 +80,6 @@
     "postcss": "^8.4.38",
     "prettier": "3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
-    "prisma": "^5.16.2",
     "tsx": "^4.16.2",
     "typescript": "^5.1.6",
     "vite": "^5.1.0",
@@ -90,8 +88,5 @@
   "engines": {
     "node": "20"
   },
-  "packageManager": "yarn@4.4.0",
-  "prisma": {
-    "seed": "tsx prisma/seed.ts"
-  }
+  "packageManager": "yarn@4.4.0"
 }
