--- prisma/schema.prisma
@@ -1,52 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "sqlite"
-  url      = env("DATABASE_URL")
-}
-
-model User {
-  id          String    @id @default(cuid())
-  email       String    @unique
-  username    String    @unique
-  name        String?
-
-  createdAt   DateTime  @default(now())
-  updatedAt   DateTime  @updatedAt
-
-  posts       Post[]
-  isAdmin     Boolean   @default(false)
-}
-
-model Post {
-  id          String    @id @default(cuid())
-  title       String
-  content     String
-  slug        String    @unique
-
-  createdAt   DateTime  @default(now())
-  updatedAt   DateTime  @updatedAt
-
-  published   Boolean   @default(false)
-  viewCount   Int       @default(0)
-  tags        Tag[]
-
-  author      User      @relation(fields: [authorId], references: [id])
-  authorId    String
-}
-
-model Tag {
-  id            String    @id @default(cuid())
-  name          String    @unique
-  description   String    @default("")
-
-  createdAt     DateTime  @default(now())
-  updatedAt     DateTime  @updatedAt
-
-  posts         Post[]
-}
